Chore: AR 관련3D 모델 및 이미지 에셋 추가
