+++ ar-nui-trainer/ar-nui-trainer/Assets.xcassets/Explore Ticket.arresourcegroup/Contents.json
@@ -1,0 +1,11 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "resources" : [
+    {
+      "filename" : "card1.arreferenceimage"
+    }
+  ]
+}
