+++ ar-nui-trainer/ar-nui-trainer/Assets.xcassets/Explore Ticket.arresourcegroup/card1.arreferenceimage/Contents.json
@@ -1,0 +1,15 @@
+{
+  "images" : [
+    {
+      "filename" : "card1.png",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "properties" : {
+    "width" : 0.059999999999999998
+  }
+}
