+++ ar-nui-trainer/ar-nui-trainer/Home/Controller/HomeViewController.swift
@@ -1,0 +1,29 @@
+//
+//  HomeViewController.swift
+//  ar-nui-trainer
+//
+//  Created by Jinyoung Yoo on 2023/07/19.
+//
+
+import UIKit
+
+class HomeViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
